gcp-services/gcs-file-ingest-service/requirements.txt
Flask>=2.2
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.12.0
pandas>=2.1.0
pyarrow>=12.0.0
gunicorn  # optional for local server testing
